requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
starlette
itsdangerous
python-multipart
psycopg2-binary>=2.9.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0